Implement resize command
