+++ src/transforms/resize.rs
@@ -1,9 +1,10 @@
-use std::arch::aarch64::vmulxs_f32;
-use image::{DynamicImage, GenericImageView, Pixel, Rgb, RgbImage};
+// use std::arch::arch64::vmulxs_f32;
+use image::{DynamicImage, Rgb};
 
 use crate::chain::{ProcessableImage, ProcessRecord, ProcessOptions, ResizeOptions};
-use crate::canvas::Canvas;
+use crate::transform_processor::{Canvas, TransformProcessor};
 use crate::image_service::Fit;
+use crate::transform_processor::resize::Resize;
 
 pub fn resize_with_pad(img_instance: &DynamicImage, resize_width: u32, resize_height: u32, bg_color: Option<Rgb<u8>>) -> DynamicImage {
     let mut img_for_processing = ProcessableImage {
@@ -20,36 +21,11 @@ pub fn resize_with_pad(img_instance: &DynamicImage, resize_width: u32, resize_he
             mode: Fit::Pad,
         }, bg_color },
     };
-    let canvas = Canvas::default();
-
-
+    let mut canvas = Canvas::new(Resize::default());
+    
     canvas.execute(&mut img_for_processing);
-
-    let mut img_canvas = img_for_processing.out_img.unwrap();
-    // let mut img_canvas = RgbImage::new(resize_width, resize_height);
-    //
-    // for x in 0..resize_width {
-    //     for y in 0..resize_height {
-    //         img_canvas.put_pixel(x, y, bg_color.unwrap_or(Rgb([255u8, 0u8, 0u8])));
-    //     }
-    // }
-
-    let img_overlay = img_instance.resize(resize_width, resize_height, image::imageops::FilterType::Lanczos3);
-    let (overlay_width, overlay_height) = img_overlay.dimensions();
-    let offset_x = (img_canvas.width() - overlay_width) / 2;
-    let offset_y = (img_canvas.height() - overlay_height) / 2;
-
-    for x in 0..overlay_width {
-        for y in 0..overlay_height {
-            if x + offset_x < overlay_width + offset_x && y + offset_y < overlay_height + offset_y {
-                let overlay_pixel = img_overlay.get_pixel(x, y);
-
-                img_canvas.put_pixel(x + offset_x, y + offset_y, overlay_pixel.to_rgb());
-            }
-        }
-    }
-
-    DynamicImage::from(img_canvas)
+    
+    DynamicImage::from(img_for_processing.out_img.unwrap())
 }
 
 #[cfg(test)]
